Trait itron_solid_asp3::task::IntoProcessorSet

pub trait IntoProcessorSet: Sealed + Sized { }

This is supported on crate features unstable and dcre only.

The trait implemented by types that can be passed to crate::task::Builder::processor_affinity. This trait is sealed.